requirements.txt
google-generativeai==0.5.0
python-dotenv==1.0.0
chromadb==0.4.22
langchain==0.1.0
PyPDF2==3.0.1
streamlit==1.32.0
python-magic==0.4.27
tiktoken==0.5.2
faiss-cpu==1.7.4 